pyproject.toml
[project]
name = "enhancer"
version = "0.1.0"
description = "Finegrain Image Enhancer"
authors = [
    { name = "Laurent Fainsin", email = "laurent@lagon.tech" }
]
dependencies = [
    "gradio>=4.42.0",
    "pillow>=10.4.0",
    "gradio-imageslider>=0.0.20",
    "pillow-heif>=0.18.0",
    "refiners @ git+https://github.com/finegrain-ai/refiners",
    "spaces>=0.29.3",
    "numpy<2.0.0",
]
readme = "README.md"
requires-python = ">= 3.12, <3.13"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/enhancer"]

[tool.ruff]
src = ["src"]  # https://docs.astral.sh/ruff/settings/#src
exclude = ["esrgan_model.py"]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = [
  "E",   # pycodestyle errors
  "W",   # pycodestyle warnings
  "F",   # pyflakes
  "UP",  # pyupgrade
  "A",   # flake8-builtins
  "B",   # flake8-bugbear
  "Q",   # flake8-quotes
  "I",   # isort
]

[tool.pyright]
include = ["src"]
exclude = ["**/__pycache__"]
